+++ dbdaora/hash/_tests/test_integration_service_hash_aioredis_get_one.py
@@ -19,6 +19,19 @@ async def test_should_get_one(
 
 
 @pytest.mark.asyncio
+async def test_should_get_one_without_cache(
+    fake_service, serialized_fake_entity, fake_entity
+):
+    fake_service.cache = None
+    await fake_service.repository.memory_data_source.hmset(
+        'fake:fake', *itertools.chain(*serialized_fake_entity.items())
+    )
+    entity = await fake_service.get_one('fake')
+
+    assert entity == fake_entity
+
+
+@pytest.mark.asyncio
 async def test_should_get_one_from_cache(
     fake_service, serialized_fake_entity, fake_entity
 ):
@@ -36,6 +49,24 @@ async def test_should_get_one_from_cache(
 async def test_should_get_one_from_fallback_after_open_circuit_breaker(
     fake_service, fake_entity, mocker
 ):
+    fake_service.repository.memory_data_source.hgetall = asynctest.CoroutineMock(
+        side_effect=RedisError
+    )
+    fake_service.repository.fallback_data_source.db[
+        'fake:fake'
+    ] = dataclasses.asdict(fake_entity)
+
+    entity = await fake_service.get_one('fake')
+
+    assert entity == fake_entity
+    assert fake_service.logger.warning.call_count == 1
+
+
+@pytest.mark.asyncio
+async def test_should_get_one_from_fallback_after_open_circuit_breaker_without_cache(
+    fake_service, fake_entity, mocker
+):
+    fake_service.cache = None
     fake_service.repository.memory_data_source.hgetall = asynctest.CoroutineMock(
         side_effect=RedisError
     )